backend/requirements.txt
# ஜோதிட AI - Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Astronomy / Ephemeris
pyswisseph>=2.10.3.2

# Data Processing
pydantic>=2.5.3
python-dateutil>=2.8.2
pytz>=2024.1

# HTTP client
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0

# Database - PostgreSQL
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
itsdangerous>=2.1.2

# PDF Report Generation
reportlab>=4.0.8

# Development (optional)
# pytest>=7.4.4
# pytest-asyncio>=0.23.3
